requirements.txt
python-dotenv>=1.0.0
requests>=2.32.0
youtube-transcript-api>=0.6.2,<1.0.0
yt-dlp>=2024.12.13
py-clob-client>=0.34.6
